+++ lib/src/internal/tabbed_view_provider.dart
@@ -1,11 +1,8 @@
 import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 import 'package:tabbed_view/src/tabbed_view.dart'
-    show
-        OnTabClose,
-        OnTabSelection,
+    show OnTabSelection,
         TabBarPosition,
-        TabCloseInterceptor,
         TabSelectInterceptor,
         TabsAreaButtonsBuilder;
 import 'package:tabbed_view/src/tabbed_view_controller.dart';
@@ -13,6 +10,7 @@ import 'package:tabbed_view/src/tabbed_view_menu_item.dart';
 import 'package:tabbed_view/src/typedefs/on_before_drop_accept.dart';
 import 'package:tabbed_view/src/typedefs/on_draggable_build.dart';
 import 'package:tabbed_view/src/typedefs/can_drop.dart';
+import 'package:tabbed_view/src/typedefs/on_tab_secondary_tap.dart';
 
 /// Propagates parameters to internal widgets.
 @internal
@@ -28,6 +26,7 @@ class TabbedViewProvider {
       required this.selectToEnableButtons,
       this.closeButtonTooltip,
       this.tabsAreaButtonsBuilder,
+      this.onTabSecondaryTap,
       required this.menuItems,
       required this.menuItemsUpdater,
       required this.onTabDrag,
@@ -36,7 +35,9 @@ class TabbedViewProvider {
       required this.canDrop,
       required this.onBeforeDropAccept,
       required this.dragScope,
-      required this.tabBarPosition});
+      required this.tabBarPosition,
+      this.hiddenTabsMenuItemBuilder,
+      this.trailing});
 
   final TabbedViewController controller;
   final bool contentClip;
@@ -45,6 +46,7 @@ class TabbedViewProvider {
   final TabCloseInterceptor? tabCloseInterceptor;
   final OnTabSelection? onTabSelection;
   final TabSelectInterceptor? tabSelectInterceptor;
+  final OnTabSecondaryTap? onTabSecondaryTap;
   final bool selectToEnableButtons;
   final String? closeButtonTooltip;
   final TabsAreaButtonsBuilder? tabsAreaButtonsBuilder;
@@ -57,6 +59,8 @@ class TabbedViewProvider {
   final OnBeforeDropAccept? onBeforeDropAccept;
   final String? dragScope;
   final TabBarPosition tabBarPosition;
+  final HiddenTabsMenuItemBuilder? hiddenTabsMenuItemBuilder;
+  final Widget? trailing;
 }
 
 /// Updater for menu items