+++ lib/src/internal/tabs_area/tabs_area_corner.dart
@@ -4,6 +4,9 @@ import 'package:tabbed_view/src/internal/tabbed_view_provider.dart';
 import 'package:tabbed_view/src/internal/tabs_area/drop_tab_widget.dart';
 import 'package:tabbed_view/src/internal/tabs_area/hidden_tabs.dart';
 import 'package:tabbed_view/src/internal/tabs_area/tabs_area_buttons_widget.dart';
+import 'package:tabbed_view/src/tabbed_view.dart';
+import 'package:tabbed_view/src/theme/tab_theme_data.dart';
+import 'package:tabbed_view/src/theme/theme_widget.dart';
 
 @internal
 class TabsAreaCorner extends StatelessWidget {
@@ -19,8 +22,29 @@ class TabsAreaCorner extends StatelessWidget {
   }
 
   Widget _builder(BuildContext context, Widget? child) {
-    Widget cornerContent =
-        TabsAreaButtonsWidget(provider: provider, hiddenTabs: hiddenTabs);
+    final List<Widget> children = [
+      TabsAreaButtonsWidget(provider: provider, hiddenTabs: hiddenTabs)
+    ];
+
+    if (provider.trailing != null) {
+      Widget trailing = provider.trailing!;
+      if (provider.tabBarPosition.isVertical) {
+        final TabThemeData tabTheme = TabbedViewTheme.of(context).tab;
+        if (tabTheme.rotateCharactersInVerticalTabs) {
+          final int quarterTurns =
+              provider.tabBarPosition == TabBarPosition.left ? -1 : 1;
+          trailing = RotatedBox(quarterTurns: quarterTurns, child: trailing);
+        }
+      }
+      children.add(trailing);
+    }
+
+    Widget cornerContent;
+    if (provider.tabBarPosition.isHorizontal) {
+      cornerContent = Row(mainAxisSize: MainAxisSize.min, children: children);
+    } else {
+      cornerContent = Column(mainAxisSize: MainAxisSize.min, children: children);
+    }
 
     Widget corner = Container(
         padding: provider.tabBarPosition.isHorizontal